+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/bbm/associations/TokenAssociationId.java
@@ -16,25 +16,14 @@
 
 package com.hedera.node.app.bbm.associations;
 
-import static com.hedera.node.app.bbm.associations.TokenAssociation.toLongsPair;
-import static com.hedera.node.app.bbm.associations.TokenAssociation.toPair;
-
 import com.google.common.collect.ComparisonChain;
 import com.hedera.node.app.bbm.utils.Writer;
-import com.hedera.node.app.service.mono.state.virtual.entities.OnDiskTokenRel;
-import com.hedera.node.app.state.merkle.disk.OnDiskKey;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
 record TokenAssociationId(long accountId, long tokenId) implements Comparable<TokenAssociationId> {
     static TokenAssociationId fromMod(@NonNull final com.hedera.hapi.node.base.TokenAssociation association) {
         return new TokenAssociationId(
                 association.accountId().accountNum(), association.tokenId().tokenNum());
-    }
-
-    static TokenAssociationId fromMono(@NonNull final OnDiskKey<OnDiskTokenRel> tokenRel) {
-        final var key = toLongsPair(toPair(tokenRel.getKey().getKey()));
-        ;
-        return new TokenAssociationId(key.left(), key.right());
     }
 
     @Override
