+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/bbm/contracts/ContractUtils.java
@@ -16,19 +16,6 @@
 
 package com.hedera.node.app.bbm.contracts;
 
-import com.hedera.node.app.service.mono.state.adapters.VirtualMapLike;
-import com.hedera.node.app.service.mono.state.migration.AccountStorageAdapter;
-import com.hedera.node.app.service.mono.state.virtual.VirtualBlobKey;
-import com.hedera.node.app.service.mono.state.virtual.VirtualBlobKey.Type;
-import com.hedera.node.app.service.mono.state.virtual.VirtualBlobValue;
-import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeSet;
-
 public class ContractUtils {
 
     static final int ESTIMATED_NUMBER_OF_CONTRACTS = 100_000;
@@ -36,70 +23,5 @@ public class ContractUtils {
 
     private ContractUtils() {
         // Utility class
-    }
-
-    /**
-     * Return all the bytecodes for all the contracts in this state.
-     */
-    @NonNull
-    public static Contracts getMonoContracts(
-            VirtualMapLike<VirtualBlobKey, VirtualBlobValue> files, AccountStorageAdapter accountAdapter) {
-        final var contractIds = getAllKnownContracts(accountAdapter);
-        final var deletedContractIds = getAllDeletedContracts(accountAdapter);
-        final var contractContents = getAllContractContents(files, contractIds, deletedContractIds);
-        return new Contracts(contractContents, deletedContractIds, contractIds.size());
-    }
-
-    /**
-     * Returns all contracts known via Hedera accounts, by their contract id (lowered to an Integer)
-     */
-    @NonNull
-    private static Set</*@NonNull*/ Integer> getAllKnownContracts(AccountStorageAdapter accounts) {
-        final var ids = new HashSet<Integer>(ESTIMATED_NUMBER_OF_CONTRACTS);
-        accounts.forEach((k, v) -> {
-            if (null != k && null != v && v.isSmartContract()) {
-                ids.add(k.intValue());
-            }
-        });
-        return ids;
-    }
-
-    /** Returns the ids of all deleted contracts ("self-destructed") */
-    @NonNull
-    private static Set</*@NonNull*/ Integer> getAllDeletedContracts(AccountStorageAdapter accounts) {
-        final var ids = new HashSet<Integer>(ESTIMATED_NUMBER_OF_DELETED_CONTRACTS);
-        accounts.forEach((k, v) -> {
-            if (null != k && null != v && v.isSmartContract() && v.isDeleted()) {
-                ids.add(k.intValue());
-            }
-        });
-        return ids;
-    }
-
-    /** Returns the bytecodes for all the requested contracts */
-    @NonNull
-    private static Collection</*@NonNull*/ Contract> getAllContractContents(
-            @NonNull final VirtualMapLike<VirtualBlobKey, VirtualBlobValue> fileStore,
-            @NonNull final Collection</*@NonNull*/ Integer> contractIds,
-            @NonNull final Collection</*@NonNull*/ Integer> deletedContractIds) {
-        Objects.requireNonNull(contractIds);
-        Objects.requireNonNull(deletedContractIds);
-
-        final var codes = new ArrayList<Contract>(ESTIMATED_NUMBER_OF_CONTRACTS);
-        for (final var cid : contractIds) {
-            final var vbk = new VirtualBlobKey(Type.CONTRACT_BYTECODE, cid);
-            if (fileStore.containsKey(vbk)) {
-                final var blob = fileStore.get(vbk);
-                if (null != blob) {
-                    final var c = new Contract(
-                            new TreeSet<>(),
-                            blob.getData(),
-                            deletedContractIds.contains(cid) ? Validity.DELETED : Validity.ACTIVE);
-                    c.ids().add(cid);
-                    codes.add(c);
-                }
-            }
-        }
-        return codes;
     }
 }
