+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/bbm/tokentypes/TokenTypesDumpUtils.java
@@ -22,19 +22,15 @@ import static com.swirlds.common.threading.manager.AdHocThreadManager.getStaticT
 import com.hedera.hapi.node.base.TokenID;
 import com.hedera.hapi.node.base.TokenSupplyType;
 import com.hedera.hapi.node.base.TokenType;
-import com.hedera.node.app.bbm.DumpCheckpoint;
 import com.hedera.node.app.bbm.utils.FieldBuilder;
 import com.hedera.node.app.bbm.utils.ThingsToStrings;
 import com.hedera.node.app.bbm.utils.Writer;
 import com.hedera.node.app.service.mono.legacy.core.jproto.JKey;
-import com.hedera.node.app.service.mono.state.adapters.MerkleMapLike;
 import com.hedera.node.app.service.mono.state.adapters.VirtualMapLike;
-import com.hedera.node.app.service.mono.state.merkle.MerkleToken;
-import com.hedera.node.app.service.mono.utils.EntityNum;
+import com.hedera.node.app.service.mono.statedumpers.DumpCheckpoint;
 import com.hedera.node.app.state.merkle.disk.OnDiskKey;
 import com.hedera.node.app.state.merkle.disk.OnDiskValue;
 import com.swirlds.base.utility.Pair;
-import com.swirlds.merkle.map.MerkleMap;
 import com.swirlds.virtualmap.VirtualMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.nio.file.Path;
@@ -99,34 +95,6 @@ public class TokenTypesDumpUtils {
             System.out.printf(
                     "=== mod tokens report is %d bytes at checkpoint %s%n", writer.getSize(), checkpoint.name());
         }
-    }
-
-    public static void dumpMonoTokenType(
-            @NonNull final Path path,
-            @NonNull final MerkleMap<EntityNum, MerkleToken> tokens,
-            @NonNull final DumpCheckpoint checkpoint) {
-        try (@NonNull final var writer = new Writer(path)) {
-            final var allTokens = gatherTokensFromMono(tokens);
-            dump(writer, allTokens);
-            System.out.printf(
-                    "=== mono tokens report is %d bytes at checkpoint %s%n", writer.getSize(), checkpoint.name());
-        }
-    }
-
-    @NonNull
-    private static Map<TokenType, Map<Long, Token>> gatherTokensFromMono(
-            @NonNull final MerkleMap<EntityNum, MerkleToken> source) {
-
-        final var allTokens = new HashMap<TokenType, Map<Long, Token>>();
-
-        allTokens.put(TokenType.FUNGIBLE_COMMON, new HashMap<>());
-        allTokens.put(TokenType.NON_FUNGIBLE_UNIQUE, new HashMap<>());
-
-        // todo check if it is possible to use multi threading with MerkleMaps like VirtualMaps
-        MerkleMapLike.from(source).forEachNode((en, mt) -> allTokens
-                .get(TokenType.fromProtobufOrdinal(mt.tokenType().ordinal()))
-                .put(en.longValue(), Token.fromMono(mt)));
-        return allTokens;
     }
 
     @NonNull
