+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/statedumpers/singleton/Congestion.java
@@ -17,10 +17,10 @@
 package com.hedera.node.app.service.mono.statedumpers.singleton;
 
 import com.hedera.hapi.node.state.throttles.ThrottleUsageSnapshot;
-import com.hedera.node.app.service.mono.statedumpers.utils.ThingsToStrings;
 import com.hedera.node.app.service.mono.pbj.PbjConverter;
 import com.hedera.node.app.service.mono.state.merkle.MerkleNetworkContext;
 import com.hedera.node.app.service.mono.state.submerkle.RichInstant;
+import com.hedera.node.app.service.mono.statedumpers.utils.ThingsToStrings;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.Arrays;
