+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/statedumpers/singleton/StakingInfoDumpUtils.java
@@ -16,11 +16,11 @@
 
 package com.hedera.node.app.service.mono.statedumpers.singleton;
 
+import com.hedera.node.app.service.mono.state.adapters.MerkleMapLike;
+import com.hedera.node.app.service.mono.state.merkle.MerkleStakingInfo;
 import com.hedera.node.app.service.mono.statedumpers.DumpCheckpoint;
 import com.hedera.node.app.service.mono.statedumpers.utils.FieldBuilder;
 import com.hedera.node.app.service.mono.statedumpers.utils.Writer;
-import com.hedera.node.app.service.mono.state.adapters.MerkleMapLike;
-import com.hedera.node.app.service.mono.state.merkle.MerkleStakingInfo;
 import com.hedera.node.app.service.mono.utils.EntityNum;
 import com.swirlds.base.utility.Pair;
 import com.swirlds.merkle.map.MerkleMap;
